hi, some test for the git in my new pc

+++ src/main/java/com/laith/hrsystem/laith/service/DirectorateService.java
@@ -3,6 +3,7 @@ package com.laith.hrsystem.laith.service;
 import com.laith.hrsystem.laith.dto.DirectorateDto;
 import com.laith.hrsystem.laith.exceptions.EmployeeNotFoundException;
 import com.laith.hrsystem.laith.exceptions.NotFoundDirectorateException;
+import com.laith.hrsystem.laith.model.Department;
 import com.laith.hrsystem.laith.model.Directorate;
 import com.laith.hrsystem.laith.repository.DirectorateRepository;
 import jakarta.transaction.Transactional;
@@ -10,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -29,29 +31,24 @@ public class DirectorateService {
         return this.directorateRepository.save(directorate);
     }
 
-
-    public Directorate updateDirectorate(DirectorateDto directorateDto) {
+    public void updateDirectorate(DirectorateDto directorateDto) {
         Directorate directorate = DirectorateDto.fromDirectorateDto(directorateDto);
-        //TODO I have to check if the department is exist or dose not
-        //  directorate.setDepartment(departmentService.findDepartmentById(directorateDto.getDepartmentId()));
-        return this.directorateRepository.save(directorate);
-    }
 
+        Department departmentById = departmentService.findDepartmentById(directorate.getDepartment().getId());
+        if(Objects.nonNull(directorate.getId() ) && Objects.nonNull(departmentById))
+        this.directorateRepository.save(directorate);
+    }
 
     public Directorate findDirectorateById(Long directorateId) {
         return directorateRepository.findById(directorateId).orElseThrow(
                 () -> new NotFoundDirectorateException("Directorate not found for this id :: " + directorateId));
-
     }
 
     public void deleteDirectorate(Long id) {
         Directorate directorate = directorateRepository.findById(id).orElseThrow(
                 () -> new EmployeeNotFoundException("" +
-                        "we can't delete this  Directorate couz it's " +
-                        " there is no employee with this id :" + id));
-
+                        "we can't delete this  Directorate because it's " + " there is no employee with this id :" + id));
+        if(Objects.nonNull(directorate))
         directorateRepository.deleteById(id);
-
     }
-
 }
